requirements.txt
Flask==3.1.0
matplotlib==3.9.3
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
scikit_learn==1.5.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
timm==1.0.12
torch==2.5.0+cu124
torchvision==0.20.0
tqdm==4.66.5
xgboost==2.1.3
